requirements.txt
abstract-rendering==0.5.1
alabaster==0.7.6
alembic==0.9.1
anaconda-client==1.1.0
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astropy==1.3.2
Babel==2.1.1
bcrypt==3.1.3
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.8.3
bleach==1.5.0
bokeh==0.10.0
boto==2.38.0
bz2file==0.98
cached-property==1.3.0
cffi==1.9.1
chainer==2.0.0
chainerrl==0.2.0
chardet==3.0.2
clyent==0.4.0
colorama==0.3.3
conda==4.3.23
conda-build==1.18.1
configobj==5.0.6
corenlp-python==3.4.1.post1
cryptography==1.7.1
cycler==0.10.0
Cython==0.23.4
cytoolz==0.7.4
dataset==0.8.0
datashape==0.5.4
decorator==4.1.2
docutils==0.12
entrypoints==0.2.3
fastcache==1.0.2
filelock==2.0.8
Flask==0.10.1
Flask-SQLAlchemy==2.3.2
future==0.16.0
gensim==2.1.0
greenlet==0.4.9
gym==0.8.2
h5py==2.7.0
html5lib==0.9999999
idna==2.2
ipykernel==4.1.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==4.1.0
itsdangerous==0.24
jdcal==1.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-contrib-core==0.3.1
jupyter-contrib-nbextensions==0.2.8
jupyter-core==4.3.0
jupyter-highlight-selected-word==0.0.11
jupyter-latex-envs==1.3.8.2
jupyter-nbextensions-configurator==0.2.5
line-bot-sdk==1.5.0
lxml==3.4.4
Mako==1.0.6
MarkupSafe==1.0
matplotlib==1.5.1
mccabe==0.6.1
mecab-python3==0.7
mistune==0.7.1
multipledispatch==0.4.8
mysqlclient==1.3.10
nb-anacondacloud==1.4.0
nb-conda==2.2.0
nb-conda-kernels==2.1.0
nbconvert==5.1.1
nbformat==4.3.0
nbpresent==3.0.2
networkx==1.10
nltk==3.1
normality==0.4.0
nose==1.3.7
notebook==4.4.1
numexpr==2.6.2
numpy==1.11.3
oauthlib==2.0.2
odo==0.3.4
openpyxl==2.2.6
pandas==0.19.2
pandocfilters==1.4.1
paramiko==2.2.1
path.py==0.0.0
patsy==0.4.1
pbr==3.0.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.0.0
ply==3.8
prompt-toolkit==1.0.15
protobuf==3.3.0
psutil==3.2.2
ptyprocess==0.5.2
py==1.4.30
pyasn1==0.2.3
pycodestyle==2.3.1
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.19.5.1
pyflakes==1.5.0
pyglet==1.2.4
Pygments==2.0.2
PyMySQL==0.7.9
PyNaCl==1.1.2
pyOpenSSL==16.2.0
pyparsing==2.0.3
PypeR==1.1.2
pytest==2.8.1
python-dateutil==2.4.2
python-editor==1.0.3
python-twitter==3.2.1
pytz==2015.6
PyWavelets==0.5.2
PyYAML==3.11
pyzmq==14.7.0
qtconsole==4.1.0
redis==2.10.3
requests==2.13.0
requests-oauthlib==0.8.0
rope-py3k==0.9.4.post1
scikit-image==0.13.0
scikit-learn==0.18.2
scipy==0.19.0
selenium==3.8.0
simplegeneric==0.8.1
six==1.10.0
slacker==0.9.60
slackweb==1.0.5
smart-open==1.5.3
snowballstemmer==1.2.0
sockjs-tornado==1.0.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spyder==2.3.7
SQLAlchemy==1.0.9
sshtunnel==0.1.2
statsmodels==0.8.0
stevedore==1.21.0
sympy==0.7.6.1
tables==3.3.0
terminado==0.5
testpath==0.3
toolz==0.7.4
tornado==4.5.1
traitlets==4.3.2
tweepy==3.5.0
ujson==1.33
unicodecsv==0.14.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.10.4
xlrd==0.9.4
XlsxWriter==0.7.7
xlwings==0.4.1
xlwt==1.0.0
xmltodict==0.10.2
